+++ Core/TimberApi/ResourceAssetSystem/ResourceAssetSystemConfigurator.cs
@@ -4,6 +4,7 @@ namespace TimberApi.ResourceAssetSystem
 {
     internal class ResourceAssetSystemConfigurator : IConfigurator
     {
+        //TODO: Add generalized patcher logic
         public void Configure(IContainerDefinition containerDefinition)
         {
             containerDefinition.Bind<AssetSystemConfiguratorPatcher>().AsSingleton();