+++ Core/TimberApi/ShaderSystem/ShaderAppliers/PathShaderApplier.cs
@@ -1,14 +1,15 @@
 ﻿using UnityEngine;
 
-namespace TimberApi.AssetShaderSystem.ShaderFix
+namespace TimberApi.ShaderSystem.ShaderAppliers
 {
-    internal class PathShaderFixApplier : IShaderFixApplier
+    internal class PathShaderApplier : IShaderApplier
     {
         private Shader _pathShader = null!;
 
         public void LoadShader()
         {
-            _pathShader = Resources.Load<GameObject>("Buildings/Paths/Path/Path.Folktails").GetComponent<MeshRenderer>().materials[0].shader;
+            _pathShader = Resources.Load<GameObject>("Buildings/Paths/Path/Path.Folktails")
+                .GetComponentInChildren<MeshRenderer>().materials[0].shader;
         }
 
         public bool ShouldApplyShader(Material material)