+++ Core/TimberApi/TimberApiBootstrapConfigurator.cs
@@ -12,7 +12,8 @@ namespace TimberApi
             containerDefinition.Bind<TimberApi>().AsSingleton();
             containerDefinition.Install(new AssetSystemGlobalConfigurator());
             containerDefinition.Install(new GolemFactionPatchConfigurator());
-            containerDefinition.Install(new ResourceAssetSystemConfigurator());
+
+            new AssetSystemConfiguratorPatcher().Load();
         }
     }
 }