+++ changelog.md
@@ -1,7 +1,11 @@
 ## Unreleased
 
 ### Changes
-- Fixed EntityLinker crashing the game ia a mod using them was disabled
+- Fixed EntityLinker crashing the game if a mod using them was disabled 
+- Fixed path shader
+- Changed placeholder versioning logic for developing
+- Fixed Timberborn ResourceLoader to work in main menu
+- Added material to shader appliers
 
 ## TimberAPI v0.5.3.1
 Hotfix